Continuing work on texturemapping

+++ Source/Balder/Rendering/Silverlight/RenderVertex.cs
@@ -75,9 +75,9 @@ namespace Balder.Rendering.Silverlight
 
 		public void MakeScreenCoordinates()
 		{
-			TranslatedScreenCoordinates.X = (int)TranslatedVector.X;
-			TranslatedScreenCoordinates.Y = (int)TranslatedVector.Y;
-			TranslatedScreenCoordinates.Z = (int)TranslatedVector.Z;
+			TranslatedScreenCoordinates.X = TranslatedVector.X;
+			TranslatedScreenCoordinates.Y = TranslatedVector.Y;
+			TranslatedScreenCoordinates.Z = TranslatedVector.Z;
 		}
 
 		public Vector TransformedVector;
